searchMatrix 하는 중

+++ src/main/java/com/vinnsmedia/crossword/dao/CrosswordDAO.java
@@ -28,9 +28,9 @@ public interface CrosswordDAO {
 	@Insert("INSERT INTO word (name, length, description, puzzle_seq) VALUES (#{name}, #{length}, #{description}, #{puzzle_seq})")
 	Boolean insertWords(Word word) throws Exception;
 	@Select("SELECT count(*) FROM word WHERE puzzle_seq=#{puzzle_seq}")
-	Integer getCount(Integer puzzle_seq) throws Exception;
+	Integer getCount(Long puzzle_seq) throws Exception;
 	@Update("UPDATE puzzle SET count=#{count} WHERE seq=#{puzzle_seq}")
-	Boolean updateCount(@Param("puzzle_seq") Integer puzzle_seq, @Param("count")Integer count) throws Exception;
+	Boolean updateCount(@Param("puzzle_seq") Long puzzle_seq, @Param("count")Integer count) throws Exception;
 	@Select("SELECT count(*) FROM puzzle WHERE title=#{title}")
 	Integer checkTitle(String title)throws Exception;
 	
@@ -43,9 +43,9 @@ public interface CrosswordDAO {
 	@Select("SELECT * FROM puzzle ORDER BY seq DESC")
 	List<CrosswordVO> findAllPuzzles()throws Exception;
 	@Select("SELECT * FROM puzzle WHERE seq=#{puzzle_seq}")
-	CrosswordVO getPuzzle(Integer puzzle_seq) throws Exception;
+	CrosswordVO getPuzzle(Long puzzle_seq) throws Exception;
 	@Select("SELECT * FROM word WHERE puzzle_seq=#{seq} ORDER BY seq ASC")
-	List<Word> getWordsByPuzzleSeq(Integer seq) throws Exception;
+	List<Word> getWordsByPuzzleSeq(Long seq) throws Exception;
 	
 	
 }
